add enough to start testing with examples
