+++ packages/@atjson/source-prism/public/app.ts
@@ -1,8 +1,11 @@
+import OffsetSource from '@atjson/offset-annotations';
+import Renderer from '@atjson/renderer-commonmark';
 import PRISMSource from '../src';
 
 document.addEventListener('paste', (evt: ClipboardEvent) => {
   let xml = evt.clipboardData.getData('text/plain');
   if (xml !== '') {
-    document.body.innerText = JSON.stringify(PRISMSource.fromRaw(xml).toJSON(), null, 2);
+    let doc = PRISMSource.fromRaw(xml).convertTo(OffsetSource);
+    document.body.innerText = new Renderer().render(doc) + '\n\n-----------------\n\n' + JSON.stringify(doc.toJSON(), null, 2);
   }
 });
