+++ packages/@atjson/source-prism/src/annotations/body.ts
@@ -1,0 +1,6 @@
+import { BlockAnnotation } from '@atjson/document';
+
+export default class Body extends BlockAnnotation {
+  static vendorPrefix = 'html';
+  static type = 'body';
+}
