+++ packages/@atjson/source-prism/src/annotations/head.ts
@@ -1,0 +1,6 @@
+import { BlockAnnotation } from '@atjson/document';
+
+export default class Head extends BlockAnnotation {
+  static vendorPrefix = 'html';
+  static type = 'head';
+}
