+++ packages/@atjson/source-prism/src/annotations/media.ts
@@ -1,0 +1,6 @@
+import { BlockAnnotation } from '@atjson/document';
+
+export default class Media extends BlockAnnotation {
+  static vendorPrefix = 'pam';
+  static type = 'body';
+}
